Made that only an admin could see list of workers and only an admin could add or remove them

+++ src/Controller/WorkersController.php
@@ -19,9 +19,14 @@ class WorkersController extends Controller {
      * @Route("/workers", name="workers")
      */
     public function index() {
-        $workers = $this->getDoctrine()->getRepository(User::class)->findBy(array('access'=>'1'));
-
-        return $this->render('workers/index.html.twig', array('workers' => $workers));
+        $auth_checker = $this->get('security.authorization_checker');
+        if($auth_checker->isGranted('ROLE_ADMIN')) {
+            $workers = $this->getDoctrine()->getRepository(User::class)->findBy(array('role'=>'ROLE_WORKER'));
+            return $this->render('workers/index.html.twig', array('workers' => $workers, 'admin' => $auth_checker->isGranted('ROLE_ADMIN')));
+        }
+        else {
+            return $this->render('workers/index.html.twig', array('admin' => $auth_checker->isGranted('ROLE_ADMIN')));
+        }
     }
 
     /**
@@ -29,47 +34,52 @@ class WorkersController extends Controller {
      * @Method({"GET", "POST"})
      */
     public function add(Request $request) {
-        $user = new User();
+        $auth_checker = $this->get('security.authorization_checker');
+        if($auth_checker->isGranted('ROLE_ADMIN')) {
+            $user = new User();
 
-        $form = $this->createFormBuilder($user)
-            ->add('email', EmailType::class, array(
-                'attr' => array('class' => 'simple-input')
-            ))
-            ->add('name', TextType::class, array(
-                'required' => false,
-                'attr' => array('class' => 'simple-input')
-            ))
-            ->add('second_name', TextType::class, array(
-                'required' => false,
-                'attr' => array('class' => 'simple-input')
-            ))
-            ->add('password', RepeatedType::class, array(
-                'type' => PasswordType::class,
-                'options' => array('attr' => array('class' => 'simple-input')),
-                'first_options'  => array('label' => 'Password'),
-                'second_options' => array('label' => 'Repeat Password'),
-            ))
-            ->add('register', SubmitType::class, array(
-                'label' => 'Add new worker',
-                'attr' => array('class' => 'modern')
-            ))
-            ->getForm();
+            $form = $this->createFormBuilder($user)
+                ->add('email', EmailType::class, array(
+                    'attr' => array('class' => 'simple-input')
+                ))
+                ->add('name', TextType::class, array(
+                    'required' => false,
+                    'attr' => array('class' => 'simple-input')
+                ))
+                ->add('second_name', TextType::class, array(
+                    'required' => false,
+                    'attr' => array('class' => 'simple-input')
+                ))
+                ->add('password', RepeatedType::class, array(
+                    'type' => PasswordType::class,
+                    'options' => array('attr' => array('class' => 'simple-input')),
+                    'first_options'  => array('label' => 'Password'),
+                    'second_options' => array('label' => 'Repeat Password'),
+                ))
+                ->add('register', SubmitType::class, array(
+                    'label' => 'Add new worker',
+                    'attr' => array('class' => 'modern')
+                ))
+                ->getForm();
 
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()){
-            $user = $form->getData();
-            $user->setAccess('1');
-            $user->setPassword(password_hash($user->getPassword(), PASSWORD_DEFAULT));
+            $form->handleRequest($request);
+            if ($form->isSubmitted() && $form->isValid()){
+                $user = $form->getData();
+                $user->setAccess('1');
+                $user->setPassword(password_hash($user->getPassword(), PASSWORD_DEFAULT));
 
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->persist($user);
-            $entityManager->flush();
-            return $this->redirectToRoute('workers');
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->persist($user);
+                $entityManager->flush();
+                return $this->redirectToRoute('workers');
+            }
+
+            return $this->render('workers/add_worker.html.twig', array(
+                'form' => $form->createView(), 'admin' => $auth_checker->isGranted('ROLE_ADMIN')));
         }
-
-        return $this->render('workers/add_worker.html.twig', array(
-            'form' => $form->createView()
-        ));
+        else {
+            return $this->render('workers/index.html.twig', array('admin' => $auth_checker->isGranted('ROLE_ADMIN')));
+        }
     }
 
     /**
@@ -77,11 +87,13 @@ class WorkersController extends Controller {
      * @Method({"GET"})
      */
     public function remove($id) {
-        $user = $this->getDoctrine()->getRepository(User::class)->find($id);
-        $entityManager = $this->getDoctrine()->getManager();
-        $entityManager->remove($user);
-        $entityManager->flush();
-
-        return $this->index();
+        $auth_checker = $this->get('security.authorization_checker');
+        if($auth_checker->isGranted('ROLE_ADMIN')) {
+            $user = $this->getDoctrine()->getRepository(User::class)->find($id);
+            $entityManager = $this->getDoctrine()->getManager();
+            $entityManager->remove($user);
+            $entityManager->flush();
+        }
+            return $this->index();
     }
 }
